Changed Sensordata Datatypes

+++ src/main/java/fi/robosailboat/webservice/boatCommunication/SensorData.java
@@ -7,11 +7,11 @@ import java.math.BigDecimal;
 @Data
 public class SensorData {
 
-    private final BigDecimal latitud;
-    private final BigDecimal longitud;
-    private final BigDecimal direction;
-    private final BigDecimal track;
+    private final double latitud;
+    private final double longitud;
+    private final double direction;
+    private final double track;
     private final int nrOfSatelites;
-    private final BigDecimal compassHeading;
+    private final double compassHeading;
 
 }
